Centralize agent results classes
Put both agent results clases in a single place so its easier to follow.
